requirements.txt
attrs==25.1.0
behave==1.2.6
certifi==2025.1.31
charset-normalizer==3.4.1
Faker==36.1.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
parse==1.20.2
parse_type==0.6.4
python-dotenv==1.0.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
